requirements.txt
python-telegram-bot==11.1.0
python-dotenv==0.15.0